+++ src/plugins/contentRenderer/index.ts
@@ -1,4 +1,4 @@
 export * from './contentRenderer.interface';
 export * from './table/tableContentRenderer.interface';
 export * from './table/body/advanced/advancedTableBodyContentRenderer.interface';
-export * from './cssDivs/cssDivsContentRenderer.interface';
+export * from './cssGrid/cssGridContentRenderer.interface';